+++ pkgs/jnigen/test/simple_package_test/java/com/github/dart_lang/jnigen/interfaces/MyConsumerRunner.java
@@ -16,26 +16,34 @@ public class MyConsumerRunner {
     this.consumer = consumer;
   }
 
+  private static class ConsumerRunnable implements Runnable {
+    private final MyConsumerRunner runner;
+    private final MyConsumer consumer;
+    private Object localArg;
+
+    public ConsumerRunnable(MyConsumerRunner runner, MyConsumer consumer, Object arg) {
+      this.runner = runner;
+      this.consumer = consumer;
+      this.localArg = arg;
+    }
+
+    @Override
+    public void run() {
+      try {
+        consumer.consume(localArg);
+      } finally {
+        localArg = null;
+        synchronized (runner) {
+          runner.finished = true;
+          runner.notifyAll();
+        }
+      }
+    }
+  }
+
   public void runOnAnotherThread(Object arg) {
     this.weakArg = new WeakReference<>(arg);
-    this.thread =
-        new Thread(
-            new Runnable() {
-              private Object localArg = arg;
-
-              @Override
-              public void run() {
-                try {
-                  consumer.consume(localArg);
-                } finally {
-                  localArg = null;
-                  synchronized (MyConsumerRunner.this) {
-                    finished = true;
-                    MyConsumerRunner.this.notifyAll();
-                  }
-                }
-              }
-            });
+    this.thread = new Thread(new ConsumerRunnable(this, consumer, arg));
     this.thread.start();
   }