+++ pkgs/hooks_runner/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.5.1-wip
+## 1.6.0-wip
 
 - Fix record_use path changing caching issue.